requirements.txt
flask
Flask-JSONRPC
tailer
gunicorn
gevent
docker-py

